Copy cc_object output files to a name that matches the module
cc_object output files match the name of the module if there are any
postprocessing steps like partial linking or prefixing symbols.  If
there are no postprocessing steps the output file matches the name
of the source file with the extension changed to ".o".  Always copy
the object to an output file that matches the module name.

Bug: 242601708
Test: TestCcObjectOutputFile
Change-Id: I086bb0d14a3c02093515f55395aa7a11473f8040