Rename sdk field tag "keep" to "ignore"
"ignore" is a better description of the effect that this tag has on the
optimization that is done as part of the sdk snapshot generation.
"keep" implies that something special is done with the field during the
optimization process but actually the optimization process completely
ignores the property, just as it does with unexported fields.

Bug: 248258460
Test: m nothing
Change-Id: Idfdfd87f927d641d4d301cbbde72b463c89155bf
(cherry picked from commit 579a69141e385f2072c1f0efcea2e207a45686e7)
Merged-In: Idfdfd87f927d641d4d301cbbde72b463c89155bf