Ignore test apexes from bp2build generated tags
The core problem I am trying to solve is making sure that stub
libraries in Bazel have a single apex available. In Soong, this will be
enforced using some graph walk which is not easy to port to Bazel.

However, we might need to revisit this when we build the enforcement
mechanism in Bazel. We likely need a `test_for` on the top level
apex_test Bazel rule so that the test apex is _allowed_ to link against impl
of the library it is trying to test.

(This CL retricts this to cc_library, I can expand this to other modules
if this is the right approach.)

Bug: 277651159
(cherry picked from https://android-review.googlesource.com/q/commit:2d72d9c424c4d450af83a349e7b4aa994ee3ce7a)
Merged-In: Iaeec22c5626df79a33785c766ed29102b1da403e
Change-Id: Iaeec22c5626df79a33785c766ed29102b1da403e