DO NOT MERGE: Changing the udc-dev version to 345990000
The udc-dev version is now lower than the prebuilts b/278817494.
Increasing it so it's higher than prebuilts but still low enough, b/231691162 suggested initially to use xx5990000 for x-dev.

I've checked tm-dev and the latest prebuilt was 3318, so 3459 prefix should be enough for udc-dev.

Change-Id: Idc85a121825aea754b258b306660ca71c6b1fdd9
BUG: 278817494
Ignore-AOSP-First: This change is needed in udc-dev only